+++ src/main/java/io/mosscloud/fibonaccinumbers/consumer/FibonacciConsumerStrategy.java
@@ -2,5 +2,5 @@ package io.mosscloud.fibonaccinumbers.consumer;
 
 import java.util.Collection;
 public interface FibonacciConsumerStrategy {
-    void consumerFibonacci(Collection<Integer> fibonacciCountResult);
+    void consumerFibonacci(Collection<Long> fibonacciCountResult);
 }
